+ Fixed a bug caused by Hibernate upgrade (0000148: Daily Work Load exception) + Added buttons to create iteration in project's iteration list + Fixed not being able to mark story as done, when all its tasks are done
git-svn-id: http://agilefant.googlecode.com/svn/trunk@4244 29784df9-99d5-7e28-3e6c-ac1830307ffc

+++ src/fi/hut/soberit/agilefant/db/hibernate/IterationDAOHibernate.java
@@ -148,7 +148,7 @@ public class IterationDAOHibernate extends GenericDAOHibernate<Iteration>
         List<Object[]> data = asList(crit);
         Map<Integer, Integer> result = new HashMap<Integer, Integer>();
         for (Object[] row : data) {
-            result.put((Integer) row[0], (Integer) row[1]);
+            result.put((Integer) row[0], ((Long) row[1]).intValue());
         }
         return result;
     }
